refact: create GameController class
