+++ src/index.ts
@@ -1,6 +1,6 @@
 import $ from "jquery";
-import Card from "./components/Card";
 import CardList from "./components/CardList";
+import GameController from "./GameController";
 import "./styles/index.scss";
 import { generateCardsArrayWithDoubleLength } from "./util/generate-cards";
 
@@ -31,35 +31,9 @@ function startGame(gameSize: number) {
 
   const cards = generateCardsArrayWithDoubleLength(gameSize);
   const cardList = new CardList(cardListEl.get(0), cards);
+  const gameController = new GameController(cardList, cards);
   cardList.render();
 
   cardListEl.show();
-  listenClickOnCardList(cardList, cards);
-}
-
-function listenClickOnCardList(cardList: CardList, cardsArr: Card[]) {
-  const cardListEl = cardList.getEl();
-  cardListEl.addEventListener("click", (e) => {
-    // Check if already has two selected cards
-    if (cardList.checkTwoSelectedCards()) {
-      return;
-    }
-
-    let el = e.target as HTMLElement;
-    if (el.tagName === "IMG") {
-      el = el.closest(".card");
-    }
-    if (el.classList.contains("card")) {
-      const cardIndex = Number(el.getAttribute("data-index"));
-      const card = cardsArr[cardIndex];
-
-      cardList.addToSelectedCards(card);
-
-      if (cardList.checkTwoSelectedCards()) {
-        const match = cardList.checkSelectedCardsMatch();
-
-        console.log(match);
-      }
-    }
-  });
+  gameController.start();
 }
